@@ -11,15 +11,17 @@ namespace ZendTest\Feed\PubSubHubbub\Model;
 
 use DateTime;
 use PDO;
+use PHPUnit\Framework\TestCase;
 use Zend\Db\Adapter\Adapter as DbAdapter;
 use Zend\Db\TableGateway\TableGateway;
 use Zend\Feed\PubSubHubbub\Model\Subscription;
+use Zend\Feed\PubSubHubbub\Model\SubscriptionPersistenceInterface;
 
 /**
  * @group      Zend_Feed
  * @group      Zend_Feed_Pubsubhubbub_Model
  */
-class SubscriptionTest extends \PHPUnit_Framework_TestCase
+class SubscriptionTest extends TestCase
 {
     /**
      * @group ZF-10069
@@ -59,9 +61,9 @@ class SubscriptionTest extends \PHPUnit_Framework_TestCase
 
     public function testImpemetsSubscriptionInterface()
     {
-        $reflection = new \ReflectionClass('Zend\Feed\PubSubHubbub\Model\Subscription');
+        $reflection = new \ReflectionClass(Subscription::class);
         $this->assertTrue(
-            $reflection->implementsInterface('Zend\Feed\PubSubHubbub\Model\SubscriptionPersistenceInterface')
+            $reflection->implementsInterface(SubscriptionPersistenceInterface::class)
         );
         unset($reflection);
     }
