@@ -10,13 +10,15 @@
 namespace ZendTest\Feed\Reader\Entry;
 
 use DateTime;
+use PHPUnit\Framework\TestCase;
 use Zend\Feed\Reader;
+use Zend\Feed\Reader\Entry\Atom;
 
 /**
 * @group Zend_Feed
 * @group Zend_Feed_Reader
 */
-class AtomStandaloneEntryTest extends \PHPUnit_Framework_TestCase
+class AtomStandaloneEntryTest extends TestCase
 {
     protected $feedSamplePath = null;
 
@@ -65,7 +67,7 @@ class AtomStandaloneEntryTest extends \PHPUnit_Framework_TestCase
         $object = Reader\Reader::importString(
             file_get_contents($this->feedSamplePath . '/id/atom10.xml')
         );
-        $this->assertInstanceOf('Zend\Feed\Reader\Entry\Atom', $object);
+        $this->assertInstanceOf(Atom::class, $object);
     }
 
     /**
