@@ -9,13 +9,15 @@
 
 namespace ZendTest\Feed\Reader\Feed;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Feed\Reader;
+use Zend\Feed\Reader\Extension\Atom\Feed;
 
 /**
 * @group Zend_Feed
 * @group Reader\Reader
 */
-class CommonTest extends \PHPUnit_Framework_TestCase
+class CommonTest extends TestCase
 {
     protected $feedSamplePath = null;
 
@@ -33,7 +35,7 @@ class CommonTest extends \PHPUnit_Framework_TestCase
         $feed = Reader\Reader::importString(
             file_get_contents($this->feedSamplePath.'/atom.xml')
         );
-        $this->assertInstanceOf('DOMDocument', $feed->getDomDocument());
+        $this->assertInstanceOf(\DOMDocument::class, $feed->getDomDocument());
     }
 
     public function testGetsDomXpathObject()
@@ -41,7 +43,7 @@ class CommonTest extends \PHPUnit_Framework_TestCase
         $feed = Reader\Reader::importString(
             file_get_contents($this->feedSamplePath.'/atom.xml')
         );
-        $this->assertInstanceOf('DOMXPath', $feed->getXpath());
+        $this->assertInstanceOf(\DOMXPath::class, $feed->getXpath());
     }
 
     public function testGetsXpathPrefixString()
@@ -57,7 +59,7 @@ class CommonTest extends \PHPUnit_Framework_TestCase
         $feed = Reader\Reader::importString(
             file_get_contents($this->feedSamplePath.'/atom.xml')
         );
-        $this->assertInstanceOf('DOMElement', $feed->getElement());
+        $this->assertInstanceOf(\DOMElement::class, $feed->getElement());
     }
 
     public function testSaveXmlOutputsXmlStringForFeed()
@@ -75,7 +77,7 @@ class CommonTest extends \PHPUnit_Framework_TestCase
         $feed = Reader\Reader::importString(
             file_get_contents($this->feedSamplePath.'/atom.xml')
         );
-        $this->assertInstanceOf('Zend\Feed\Reader\Extension\Atom\Feed', $feed->getExtension('Atom'));
+        $this->assertInstanceOf(Feed::class, $feed->getExtension('Atom'));
     }
 
     public function testReturnsNullIfExtensionDoesNotExist()
