@@ -10,16 +10,19 @@
 namespace ZendTest\Feed\Writer\Renderer\Feed;
 
 use DateTime;
+use PHPUnit\Framework\TestCase;
 use Zend\Feed\Writer;
+use Zend\Feed\Writer\Exception\ExceptionInterface;
 use Zend\Feed\Writer\Renderer;
 use Zend\Feed\Writer\Version;
 use Zend\Feed\Reader;
+use Zend\Feed\Writer\Feed;
 
 /**
  * @group      Zend_Feed
  * @group      Zend_Feed_Writer
  */
-class RssTest extends \PHPUnit_Framework_TestCase
+class RssTest extends TestCase
 {
     protected $validWriter = null;
 
@@ -42,7 +45,7 @@ class RssTest extends \PHPUnit_Framework_TestCase
     {
         $writer = new Writer\Feed;
         $feed   = new Renderer\Feed\Rss($writer);
-        $this->assertInstanceOf('Zend\Feed\Writer\Feed', $feed->getDataContainer());
+        $this->assertInstanceOf(Feed::class, $feed->getDataContainer());
     }
 
     public function testBuildMethodRunsMinimalWriterContainerProperlyBeforeICheckRssCompliance()
@@ -76,11 +79,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('This is a test feed.', $feed->getTitle());
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testFeedTitleIfMissingThrowsException()
     {
+        $this->expectException(ExceptionInterface::class);
         $rssFeed = new Renderer\Feed\Rss($this->validWriter);
         $this->validWriter->remove('title');
         $rssFeed->render();
@@ -106,11 +107,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('This is a test description.', $feed->getDescription());
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testFeedDescriptionThrowsExceptionIfMissing()
     {
+        $this->expectException(ExceptionInterface::class);
         $rssFeed = new Renderer\Feed\Rss($this->validWriter);
         $this->validWriter->remove('description');
         $rssFeed->render();
@@ -212,11 +211,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('http://www.example.com', $feed->getLink());
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testFeedLinkToHtmlVersionOfFeedIfMissingThrowsException()
     {
+        $this->expectException(ExceptionInterface::class);
         $rssFeed = new Renderer\Feed\Rss($this->validWriter);
         $this->validWriter->remove('link');
         $rssFeed->render();
@@ -416,11 +413,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($expected, $feed->getImage());
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testImageThrowsExceptionOnMissingLink()
     {
+        $this->expectException(ExceptionInterface::class);
         $this->validWriter->setImage([
                                            'uri'   => 'http://www.example.com/logo.gif',
                                            'title' => 'Image ALT'
@@ -429,11 +424,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $rssFeed->render();
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testImageThrowsExceptionOnMissingTitle()
     {
+        $this->expectException(ExceptionInterface::class);
         $this->validWriter->setImage([
                                            'uri'  => 'http://www.example.com/logo.gif',
                                            'link' => 'http://www.example.com'
@@ -442,11 +435,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $rssFeed->render();
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testImageThrowsExceptionOnMissingUri()
     {
+        $this->expectException(ExceptionInterface::class);
         $this->validWriter->setImage([
                                            'link'  => 'http://www.example.com',
                                            'title' => 'Image ALT'
@@ -455,11 +446,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $rssFeed->render();
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testImageThrowsExceptionIfOptionalDescriptionInvalid()
     {
+        $this->expectException(ExceptionInterface::class);
         $this->validWriter->setImage([
                                            'uri'         => 'http://www.example.com/logo.gif',
                                            'link'        => 'http://www.example.com',
@@ -470,11 +459,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $rssFeed->render();
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testImageThrowsExceptionIfOptionalDescriptionEmpty()
     {
+        $this->expectException(ExceptionInterface::class);
         $this->validWriter->setImage([
                                            'uri'         => 'http://www.example.com/logo.gif',
                                            'link'        => 'http://www.example.com',
@@ -485,11 +472,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $rssFeed->render();
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testImageThrowsExceptionIfOptionalHeightNotAnInteger()
     {
+        $this->expectException(ExceptionInterface::class);
         $this->validWriter->setImage([
                                            'uri'    => 'http://www.example.com/logo.gif',
                                            'link'   => 'http://www.example.com',
@@ -501,11 +486,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $rssFeed->render();
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testImageThrowsExceptionIfOptionalHeightEmpty()
     {
+        $this->expectException(ExceptionInterface::class);
         $this->validWriter->setImage([
                                            'uri'    => 'http://www.example.com/logo.gif',
                                            'link'   => 'http://www.example.com',
@@ -517,11 +500,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $rssFeed->render();
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testImageThrowsExceptionIfOptionalHeightGreaterThan400()
     {
+        $this->expectException(ExceptionInterface::class);
         $this->validWriter->setImage([
                                            'uri'    => 'http://www.example.com/logo.gif',
                                            'link'   => 'http://www.example.com',
@@ -533,11 +514,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $rssFeed->render();
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testImageThrowsExceptionIfOptionalWidthNotAnInteger()
     {
+        $this->expectException(ExceptionInterface::class);
         $this->validWriter->setImage([
                                            'uri'    => 'http://www.example.com/logo.gif',
                                            'link'   => 'http://www.example.com',
@@ -549,11 +528,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $rssFeed->render();
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testImageThrowsExceptionIfOptionalWidthEmpty()
     {
+        $this->expectException(ExceptionInterface::class);
         $this->validWriter->setImage([
                                            'uri'    => 'http://www.example.com/logo.gif',
                                            'link'   => 'http://www.example.com',
@@ -565,11 +542,9 @@ class RssTest extends \PHPUnit_Framework_TestCase
         $rssFeed->render();
     }
 
-    /**
-     * @expectedException Zend\Feed\Writer\Exception\ExceptionInterface
-     */
     public function testImageThrowsExceptionIfOptionalWidthGreaterThan144()
     {
+        $this->expectException(ExceptionInterface::class);
         $this->validWriter->setImage([
                                            'uri'    => 'http://www.example.com/logo.gif',
                                            'link'   => 'http://www.example.com',
